Split browser hit testing helpers
